+++ src/common/core/antibot/antibot.lua
@@ -105,6 +105,8 @@ function antibot:header()
 				hdr["Permissions-Policy"] = #updated > 0 and updated or nil
 			end
 		end
+	elseif self.session_data.type == "mcaptcha" then
+		csp_directives["frame-src"] = self.variables["ANTIBOT_MCAPTCHA_URL"]
 	end
 	local csp_content = ""
 	for directive, value in pairs(csp_directives) do
@@ -324,6 +326,12 @@ function antibot:display_challenge()
 		template_vars.turnstile_sitekey = self.variables["ANTIBOT_TURNSTILE_SITEKEY"]
 	end
 
+	-- mCaptcha case
+	if self.session_data.type == "mcaptcha" then
+		template_vars.mcaptcha_sitekey = self.variables["ANTIBOT_MCAPTCHA_SITEKEY"]
+		template_vars.mcaptcha_url = self.variables["ANTIBOT_MCAPTCHA_URL"]
+	end
+
 	-- Render content
 	render(self.session_data.type .. ".html", template_vars)
 
@@ -494,6 +502,47 @@ function antibot:check_challenge()
 			return nil, "error while decoding JSON from Turnstile API : " .. tdata, nil
 		end
 		if not tdata.success then
+			return false, "client failed challenge", nil
+		end
+		self.session_data.resolved = true
+		self.session_data.time_valid = now()
+		self:set_session_data()
+		return true, "resolved", self.session_data.original_uri
+	end
+
+	-- mCaptcha case
+	if self.session_data.type == "mcaptcha" then
+		read_body()
+		local args, err = get_post_args(1)
+		if err == "truncated" or not args or not args["mcaptcha__token"] then
+			return nil, "missing challenge arg", nil
+		end
+		local httpc, err = http_new()
+		if not httpc then
+			return nil, "can't instantiate http object : " .. err, nil, nil
+		end
+		local payload = {
+			token = args["mcaptcha__token"],
+			key = self.variables["ANTIBOT_MCAPTCHA_SITEKEY"],
+			secret = self.variables["ANTIBOT_MCAPTCHA_SECRET"],
+		}
+		local json_payload = cjson.encode(payload)
+		local res, err = httpc:request_uri(self.variables["ANTIBOT_MCAPTCHA_URL"] .. "/api/v1/pow/siteverify", {
+			method = "POST",
+			body = json_payload,
+			headers = {
+				["Content-Type"] = "application/json",
+			},
+		})
+		httpc:close()
+		if not res then
+			return nil, "can't send request to mCaptcha API : " .. err, nil
+		end
+		local ok, mdata = pcall(decode, res.body)
+		if not ok then
+			return nil, "error while decoding JSON from mCaptcha API : " .. mdata, nil
+		end
+		if not mdata.valid then
 			return false, "client failed challenge", nil
 		end
 		self.session_data.resolved = true